requirements.txt
asgiref==3.10.0
Django==5.2.7
djangorestframework==3.16.1
psycopg==3.2.12
psycopg2==2.9.11
sqlparse==0.5.3
tzdata==2025.2
